Split out test task
+ Move test task to separate file.
+ Add missing requires.
+ Fix path references.
